modification

+++ 4.program.c
@@ -1,11 +1,11 @@
-//Example for address of a variable
+//Example for address of a variable & type casting
 
 #include <stdio.h>
 int a=5;
 
 int main()
 {
-unsigned long int addressofA=&a;
+unsigned long int addressofA=(unsigned long int)&a; // tupe casting
 printf("The address of a is %p \n",&a);
 printf("The address of a %ld  \n",addressofA);
 return 0;
